Upgrade Object Pool System
add abstraction into object pool
