+++ Assets/com.gdcbd.bossbattle/Scripts/Components/NEW OBJECT POOL.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f0d6dabfc2af94843bfcda9ee67e2d67
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
