+++ Assets/com.gdcbd.bossbattle/Scripts/Components/PoolObject.cs.meta
