+++ Assets/com.gdcbd.bossbattle/Scripts/Components/ProjectileController.cs
@@ -18,7 +18,7 @@ namespace com.gdcbd.bossbattle.components
         {
             if (_projectilePrefab != null)
             {
-                GameObject projectile = ProjectileSpawner.Instance.Pool.Get(); //ObjectPoolManager.Instance.GetPooledObject(_projectilePrefab);
+                GameObject projectile = ProjectileSpawner.Instance.Pool.Get();  
                 
                 if (projectile != null)
                 {
