--- Assets/com.gdcbd.bossbattle/Scripts/Managers/ProjectileSpawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6c2e5d109e85c4e24b0cfa81397bc4e2
